nndet: csrc: make iou_threshold const in nms_cuda function
When building with CUDA 12.8, we encountered the undefined symbol  error
_Z8nms_cudaRKN2at6TensorES2_f. The demangled  symbol  requires  a  const
float parameter, while the implementation used a non-const  float.  This
mismatch in function signatures prevented proper  linking  of  the  CUDA
extension.

This  change  modifies  both  the  declaration  (in   cpu/nms.cpp)   and
implementation  (in  cuda/nms.cu)  to  use   'const   float'   for   the
iou_threshold parameter.

Signed-off-by: infus38 <infus38@gmail.com>