scripts, nndet: conf: deprecate gpus setting in favor of devices
Recent changes in Lightning 2.x deprecate the gpus parameter in favor
of devices, and now require explicitly specifying the accelerator.

The devices parameter can be either an integer indicating the number of
devices (such as GPUs or TPUs) to use, or a list of device IDs (e.g., [0,1])
to specify exact devices.
The accelerator parameter explicitly defines the type of hardware accelerator
to use, with options like 'gpu', 'cpu', 'tpu' or 'auto' to let Lightning
choose the best option automatically

Signed-off-by: infus38 <infus38@gmail.com>