dockerfile: remove overreaching and irrelevant licensing
Licensing should be strictly confined to the nndet Python package, as it
represents the sole intellectual contribution of this work.  Applying  a
license to the Dockerfile or its contents is not  only  unnecessary  but
also misleading, as these elements are largely derived from  third-party
sources and do not constitute original work.

Such  overreach  dilutes  the  focus  and  undermines  proper  licensing
practices. This change rectifies the misapplication and ensures clarity.

Signed-off-by: infus38 <infus38@gmail.com>